Typescript - Refactor code base
