+++ src/app/activity-detection.ts
