+++ src/app/auto-launch-controller.ts
