+++ src/app/reports-handler.ts
@@ -1,11 +1,64 @@
+import * as archiver from 'archiver';
 import { app, BrowserWindow, dialog, shell } from 'electron';
-import * as fs from 'fs';
-
 import * as electron from 'electron';
+import * as fs from 'fs';
+import * as path from 'path';
+
 import { isMac } from '../common/env';
 import { i18n } from '../common/i18n';
-import { generateArchiveForDirectory } from '../common/utils';
 
+/**
+ * Archives files in the source directory
+ * that matches the given file extension
+ *
+ * @param source {String} source path
+ * @param destination {String} destination path
+ * @param fileExtensions {Array} array of file ext
+ * @return {Promise<void>}
+ */
+const generateArchiveForDirectory = (source: string, destination: string, fileExtensions: string[]): Promise<void> => {
+
+    return new Promise((resolve, reject) => {
+        const output = fs.createWriteStream(destination);
+        const archive = archiver('zip', { zlib: { level: 9 } });
+
+        output.on('close', () => {
+            return resolve();
+        });
+
+        archive.on('error', (err) => {
+            return reject(err);
+        });
+
+        archive.pipe(output);
+
+        const files = fs.readdirSync(source);
+        files
+            .filter((file) => fileExtensions.indexOf(path.extname(file)) !== -1)
+            .forEach((file) => {
+                switch (path.extname(file)) {
+                    case '.log':
+                        archive.file(source + '/' + file, { name: 'logs/' + file });
+                        break;
+                    case '.dmp':
+                    case '.txt': // on Windows .txt files will be created as part of crash dump
+                        archive.file(source + '/' + file, { name: 'crashes/' + file });
+                        break;
+                    default:
+                        break;
+                }
+            });
+
+        archive.finalize();
+    });
+};
+
+/**
+ * Compress and export logs stored under system log directory
+ *
+ * MacOS - /Library/Logs/Symphony/
+ * Windows - AppData\Roaming\Symphony\logs
+ */
 export const exportLogs = (): void => {
     const FILE_EXTENSIONS = [ '.log' ];
     const MAC_LOGS_PATH = '/Library/Logs/Symphony/';
@@ -42,6 +95,9 @@ export const exportLogs = (): void => {
         });
 };
 
+/**
+ * Compress and export crash dump stored under system crashes directory
+ */
 export const exportCrashDumps = (): void => {
     const FILE_EXTENSIONS = isMac ? [ '.dmp' ] : [ '.dmp', '.txt' ];
     const crashesDirectory = (electron.crashReporter as any).getCrashesDirectory();
