+++ src/app/screen-snippet-handler.ts
