+++ src/app/spell-check-handler.ts
