+++ src/app/window-utils.ts
