+++ src/common/i18n-preload.ts
@@ -1,4 +1,4 @@
-import { formatString } from './format-string';
+import { formatString } from './utils';
 
 const localeCodeRegex = /^([a-z]{2})-([A-Z]{2})$/;
